fix x64 warnings

+++ src/core/FileOperationProgress.cpp
@@ -326,7 +326,7 @@ uintptr_t TFileOperationProgressType::AdjustToCPSLimit(
 //---------------------------------------------------------------------------
 uintptr_t TFileOperationProgressType::LocalBlockSize()
 {
-  int64_t Result = TRANSFER_BUF_SIZE;
+  uintptr_t Result = TRANSFER_BUF_SIZE;
   if (LocallyUsed + Result > LocalSize)
   {
     Result = static_cast<uintptr_t>(LocalSize - LocallyUsed);
